% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polArverse.R
\docType{package}
\name{polArverse}
\alias{polArverse}
\title{\code{polArverse} meta package}
\description{
Caja de Herramientas para el Análisis Poítico de Argentina.
See the README on
\href{https://github.com/PoliticaArgentina/polArverse/README.md}{Github}
}
