% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{polArverse_packages}
\alias{polArverse_packages}
\title{List all packages in  polArverse}
\usage{
polArverse_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include polArverse in the list?}
}
\description{
List all packages in  polArverse
}
\examples{
polArverse_packages()
}
